requirements.txt
torch==1.4.0
torchvision
matplotlib
pillow
